+++ CHANGELOG.md
@@ -18,6 +18,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Add filter by attribute master type.
     * Add `predicateForAttribute` function for customization filter.
     * Add filter for each columns.
+* ObjectlistviewComponent:
+    * Add support locales for captions models of projection.
+* ModelFunctions:
+    * Add `getValueFromLocales` function.
 
 ### Changed
 * `flexberry-menu`:
@@ -34,6 +38,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Fixed sorting by clicking on table header of `flexberry-objectlistview` or `flexberry-groupedit` component in Firefox.
 * Fixed blueprints dependency for generation of application prototype for hide the sidebar with sitemap after click on sitemap-node.
 * Fixed `flexberry-menu` for configuration of columns settings in `flexberry-objectlistview` component after changing current locale for internationalization to another language and back again.
+* Now `displayAttributeName` is required for only autocomplete and dropdown modes in `flexberry-lookup`.
 
 ### Known issues
 * TransitionAborted error get thrown when adding query params (callback is called once for each query param you set to refreshModel: true).